Ryhmänomistajan autentikointi

+++ backend/models/group.js
@@ -2,9 +2,9 @@
 
 import { pool } from "../helpers/dbHelper.js"
 
-const queryPostGroup = async (fk_ownerid, groupname, groupdescription) => {
+const queryPostGroup = async (ownerid, groupname, groupdescription) => {
     return await pool.query(`INSERT INTO "groups" (fk_ownerid, groupname, groupdescription) VALUES ($1, $2, $3) RETURNING *`,
-        [fk_ownerid, groupname, groupdescription]
+        [ownerid, groupname, groupdescription]
     )
 }
 
@@ -16,15 +16,15 @@ const queryGroupById = async (id) => {
     return await pool.query(`SELECT * FROM "groups" WHERE groupid = $1`, [id])
 }
 
-const queryUpdateGroup = async (groupid, fk_ownerid, groupname, groupdescription)=> {
+const queryUpdateGroup = async (groupid, ownerid, groupname, groupdescription)=> {
     return await pool.query(`UPDATE "groups"
 SET fk_ownerid = $2, groupname = $3, groupdescription = $4
-WHERE groupid = $1`, [groupid, fk_ownerid, groupname, groupdescription]
+WHERE groupid = $1 RETURNING *`, [groupid, ownerid, groupname, groupdescription]
     )
 }
 
-const queryDeleteGroup = async (groupid) => {
-    return await pool.query(`DELETE FROM "groups" WHERE groupid = $1`, [groupid])
+const queryDeleteGroup = async (groupid, ownerid) => {
+    return await pool.query(`DELETE FROM groups WHERE groupid=$1 AND fk_ownerid=$2 RETURNING *`, [groupid, ownerid])
 }
 
 export { queryAllGroups, queryGroupById, queryPostGroup, queryDeleteGroup, queryUpdateGroup }
